requirements.txt
fastapi==0.115.12
uvicorn
gunicorn
pydantic
python-multipart
python-decouple
# chromadb
faiss-cpu
redis
openai
# numpy